refactor : Extract Clue & Map component
