+++ treasureHunt/css/currentWaypointStyles.js
@@ -49,13 +49,9 @@ const currentWaypointStyles = StyleSheet.create({
 		flex: 1,
 		display: 'flex',
 		flexDirection: 'column',
-		flexWrap: 'nowrap',
-		justifyContent: 'space-around',
-		alignItems: 'center',
+		justifyContent: 'center',
 		marginTop: 50,
 		marginBottom: 30,
-		paddingBottom: 90,
-		paddingTop: 50,
 		position: 'relative',
 		width: screenWidth - 40,
 		borderWidth: 2,
@@ -63,90 +59,32 @@ const currentWaypointStyles = StyleSheet.create({
 		borderRadius: 25,
 	},
 	mapData: {
-		display: 'flex',
-		position: 'relative',
-		flexDirection: 'column',
-		justifyContent: 'space-between',
+		flex: 1,
 		alignItems: 'center',
-		width: screenWidth - 20,
-		height: screenHeight - 40,
-		paddingTop: 10,
-		paddingBottom: 10,
-		paddingLeft: 10,
-		paddingRight: 10,
-		marginLeft: 10,
-	},
-	mapData2: {
-		display: 'flex',
-		position: 'relative',
-		flexDirection: 'column',
-		justifyContent: 'space-between',
-		alignItems: 'center',
-		width: screenWidth - 20,
-		height: screenHeight - 40,
-		paddingTop: 90,
-		paddingBottom: 10,
-		paddingLeft: 10,
-		paddingRight: 10,
-		marginLeft: 10,
-	},
-
-	button: {
-		position: 'absolute',
-		left: 0,
-		bottom: 50,
-		zIndex: 999,
+		top: 10,
 	},
 	acornContainer: {
 		position: 'absolute',
-		bottom: 110,
-		right: 35,
+		bottom: 50,
+		right: 20,
 		flexDirection: 'column-reverse',
 	},
 	acorn: {
 		height: 40,
 		width: 40,
 		marginTop: 5,
-		borderRadius: 10,
 	},
 	smallTxt: {
 		fontSize: 20,
-		fontWeight: '400',
-		paddingLeft: 10,
+		fontWeight: '600',
 		lineHeight: 20,
 		color: '#fff',
 	},
 	mapBtn: {
-		display: 'flex',
-		margin: 10,
-		left: 10,
-		bottom: 10,
-		borderRadius: 10,
-		width: 150,
-		position: 'relative',
-		height: 40,
-		alignItems: 'center',
-		justifyContent: 'flex-end',
-	},
-	woodenBtn: {
-		position: 'absolute',
-		top: 0,
-		left: 0,
-		right: 0,
-		bottom: 0,
+		bottom: -100,
+		flex: 1,
 		justifyContent: 'center',
 		alignItems: 'center',
-	},
-	mapBtnText: {
-		display: 'flex',
-		position: 'absolute',
-		padding: 'auto',
-		margin: 'auto',
-		fontWeight: '700',
-		flexDirection: 'row',
-		alignItems: 'center',
-		fontSize: 15,
-		color: '#fff',
 	},
 	clueSwipeInstruction: {
 		position: 'absolute',
